rename files and folders to match package names
